+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+3.5
+---
+...
+
 3.4
 ---
 - Add "start" annotation to event rate chart, showing each time a correlator process was started or restarted.